(fix(fixing register and login api structure) fixing user controller naming fixing user router naming fixing article controller structure fixing query controller structure
adding article model
[finish(naming controller,router)]
