+++ src/routes/index.js
@@ -1,9 +1,11 @@
 import express from 'express'
-import Register from '../controllers/register.js'
+import userRegisterController from '../controllers/user/userRegisterController.js'
+import userAuthController from '../controllers/user/userAuthController.js'
 var router = express.Router()
 
-router.post('/register',Register.postUser)
-router.post('/auth',Register.loginUser)
+router.post('/register',userRegisterController.postUser)
+router.post('/register/admin',userRegisterController.postAdmin)
+router.post('/auth',userAuthController.loginUser)
 
 
 export default router
